added enemies, goal, auto targeting

+++ Assets/Scripts/IDamageable.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(float amount);
+    void TakeDamage(int amount);
 
     void OnDeath();
 }
